refactor: remove devtools barrel files and update direct import paths for devtool loaders
