--- web/app/components/devtools/index.ts
@@ -1,2 +1,0 @@
-export * from './react-scan'
-export * from './tanstack'
