--- web/app/components/devtools/react-scan/index.ts
@@ -1,2 +1,0 @@
-export { ReactScanLoader } from './loader'
-export { ReactScan } from './scan'
