--- web/app/components/devtools/tanstack/index.ts
@@ -1,2 +1,0 @@
-export { TanStackDevtoolsWrapper } from './devtools'
-export { TanStackDevtoolsLoader } from './loader'
