+++ web/context/query-client.tsx
@@ -2,7 +2,7 @@
 
 import type { FC, PropsWithChildren } from 'react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { TanStackDevtoolsLoader } from '@/app/components/devtools'
+import { TanStackDevtoolsLoader } from '@/app/components/devtools/tanstack/loader'
 
 const STALE_TIME = 1000 * 60 * 30 // 30 minutes
